break: move awos_ingest out of src tree
